requirements.txt
requests
langchain
pypdf2
langgraph
python-dotenv
langchain-google-genai
streamlit
